requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-common==1.1.28
azure-core==1.30.1
azure-functions==1.20.0
azurefunctions-extensions-http-fastapi==1.0.0b1
azure-identity==1.16.0
azure-search-documents==11.6.0b3
fastapi==0.110.1
langchain==0.1.17
langchain-community==0.0.37
langchain-core==0.1.52
langchain-openai==0.0.2.post1
langchain-text-splitters==0.0.1
openai==1.7.2
pydantic==2.7.1
pypdf==4.0.1
pytest==8.2.0
python-dotenv==1.0.1
starlette==0.37.2
tiktoken==0.5.2
uvicorn==0.29.0